requirements.txt
flask==3.0.0
numpy==1.24.3
pandas==2.0.3
gunicorn==21.2.0
pymongo==4.6.1
python-dotenv==1.0.0